Code before lecture 4: All the classes we need for 3D graphics!
That includes a Color class, a Vector class, a Matrix class, some
utility functions, and a few empty files we're going to fill in.

Also: the code in main now uses the SDL interface abstraction in
sdl.h/cpp and we're rendering procedurally (just a full-screeen
gradient at the moment).