Don't update the screen on each rendered row (could be quite slow).
On simple scenes it may turn out that a significant portion of the
time is spent in the rendering feedback: during the lecture, we
said to redraw the screen after each rendered row, using
updateVFB().
While rendering feedback is important, it's not good if it wastes
most of the resources; the new code limits the update frequency
to at most 10 times per second.