Update .CBP files with the added sources and EXR libs.
This is not very user friendly right now, as we also require the
OpenEXR library. OpenEXR is easily installed on Linux (search for
`OpenEXR-devel' or `libopenexr-dev' in your package manager), but
for Windows, a simpler approach to automate SDK download/install
with a script is now underway, to be published later.