Fix bugs in Lamber and Phong shaders.
The bug was causing the black lines near the edges of the soccer ball.
The cause of the bug was an improper use of the faceforward() function
when computing the lambert coefficient. The normal vector was oriented
so it matches up with the camera ray. For lambert computation, though,
the dot product is taken between the normal and the LIGHT ray, so
faceforward should have been

Vector v1 = -faceforward(v2, info.normal).

A simpler code is commited instead that inverts the v2 vector and retains
the v1 (proper normal) vector, to be used in Phong's call to reflect()
later.