Fix crashes in interactive mode (missing render locks).
The render lock we were using in SDL function was not really created.
Instead of crashing on trying to acquire a NULL lock, SDL apparently
goes on... but that causes other crashes as multiple threads try to
update the screen at once.

Also add a marking function, which paints the screen in yellow by mask.
Used to show which pixels need anti-aliasing.